pyproject.toml
[project]
name = "supermileage-pi-server"
version = "0.1.0"
description = "Raspberry Pi server for supermileage vehicles"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.10.11",
    "asyncpg>=0.30.0",
    "nest-asyncio>=1.6.0",
    "psycopg2-binary>=2.9.10",
    "pyserial>=3.5",
    "python-socketio>=5.13.0",
    "dotenv>=0.9.9",
]

[dependency-groups]
dev = [
    "coverage>=7.6.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["F", "I"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["test"]
pythonpath = ["./src"]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test/*",
    "*/__pycache__/*",
]
